requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai==1.35.3
python-dotenv==1.0.1
google-auth==2.29.0
google-api-python-client==2.126.0
